+++ CHANGELOG.md
@@ -7,4 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+- [2024-04-28] Add build script that bundles the file into build/index.sh
+
 ## [0.1.0] - 2024-04-28 (The Initial Update)